testing: consolidated BO tests
